+++ src/main/resources/templates/test.html
@@ -20,14 +20,10 @@
     $(document).ready(function() {
         let fileCallPath = '';
 
-        function show(result){
-            fileCallPath = encodeURIComponent(result.path+"/s_"+result.uuid+"_"+result.name);
+        function show(fileCallPath){
+            console.log(fileCallPath);
             document.querySelector("img").src = '/displayPostThumbnail?fileName='+fileCallPath;
         }
-
-        $.getJSON("/getPostThumbnail?postId=1" , function(result) {
-            show(result);
-        });
 
         $("#uploadBtn").on("click", () => {
             var formData = new FormData();
@@ -36,7 +32,6 @@
             var file = inputFile[0].files[0];
 
             formData.append("uploadFile", file);
-            formData.append("postId",1);
 
             $.ajax({//호출
                 url : '/uploadPostThumbnail',
@@ -46,18 +41,21 @@
                 type : 'POST',
                 dataType: 'json',
                 success : function(result) {
-                    show(result);
+                    fileCallPath = encodeURIComponent(result.path+"/s_"+result.uuid+"_"+result.name);
+                    show(fileCallPath);
                 },
             });
 
         });
-        $("#removeBtn").on("click",() => {// 지우기버튼을 클릭하면 컨트롤러 내부적으로 서버의 파일과, db데이터도 지운다.
+        $("#removeBtn").on("click",() => {
             $.ajax({
                 url: '/deletePostThumbnail',
-                data: {fileName: fileCallPath, postId: 1},
+                data: {fileName: fileCallPath},
                 dataType:'text',
                 type: 'POST',
                 success: function(result){
+                    fileCallPath = '';
+                    show(fileCallPath)
                     alert(result);
                 }
             }); //$.ajax
